+++ packages/pedometer/CHANGELOG.md
@@ -1,3 +1,3 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## 1.1.0
+* The Pedometer stream now returns the value since the plugin was started on both platforms. 
+* Previously this was only the case on iOS, and Android returned the steps taken since last phone boot-up.